refactor(node): split PeerManager helpers
Extract peer persistence, routing, status, and messaging helpers.
Update node and HTTP call sites to use new accessors.
Add PeerAlertCoordinator test coverage.
