+++ src/main/java/network/crypta/clients/http/DarknetConnectionsToadlet.java
@@ -364,7 +364,7 @@ public class DarknetConnectionsToadlet extends ConnectionsToadlet {
    */
   @Override
   protected PeerNodeStatus[] getPeerNodeStatuses(boolean noHeavy) {
-    return node.getPeers().getDarknetPeerNodeStatuses(noHeavy);
+    return node.getPeers().statusBook().getDarknetPeerNodeStatuses(noHeavy);
   }
 
   /**
