+++ src/main/java/network/crypta/io/comm/IncomingPacketFilterImpl.java
@@ -123,7 +123,7 @@ public class IncomingPacketFilterImpl implements IncomingPacketFilter {
     // Mix a small amount of arrival-time entropy into the node RNG. The bias preserves legacy
     // behavior and balances entropy quality with performance.
     node.getRandom().acceptTimerEntropy(fnpTimingSource, 0.25);
-    PeerNode opn = node.getPeers().getByPeer(peer, mangler);
+    PeerNode opn = node.getPeers().roster().getByPeer(peer, mangler);
 
     if (opn == null) {
       LOG.info("Got packet from unknown address");
