+++ src/main/java/network/crypta/node/CHKInsertSender.java
@@ -619,6 +619,7 @@ public final class CHKInsertSender extends BaseSender
 
   private PeerNode findNextPeer() {
     return node.getPeers()
+        .routingSelector()
         .closerPeer(
             forkedRequestTag == null ? source : null,
             nodesRoutedTo,
