+++ src/main/java/network/crypta/node/simulator/TestUtil.java
@@ -43,7 +43,7 @@ public class TestUtil {
     while (seconds < 600) {
       Thread.sleep(1000);
       int seeds = node.getPeers().countSeednodes();
-      int seedConns = node.getPeers().getConnectedSeedServerPeersVector(null).size();
+      int seedConns = node.getPeers().seedPeers().getConnectedSeedServerPeersVector(null).size();
       int opennetPeers = node.getPeers().countValidPeers();
       int opennetConns = node.getPeers().countConnectedOpennetPeers();
       LOG.info(
